Fix interaction prediction issues
