+++ Content.Shared/Containers/ItemSlot/ItemSlotsSystem.cs
@@ -180,7 +180,7 @@ namespace Content.Shared.Containers.ItemSlots
                 if (slot.Item != null)
                     _handsSystem.TryPickupAnyHand(args.User, slot.Item.Value, handsComp: hands);
 
-                Insert(uid, slot, args.Used, args.User, excludeUserAudio: args.Predicted);
+                Insert(uid, slot, args.Used, args.User, excludeUserAudio: true);
                 args.Handled = true;
                 return;
             }
