+++ Content.Shared/Emag/Components/EmagComponent.cs
@@ -3,13 +3,13 @@ namespace Content.Shared.Emag.Components
     [RegisterComponent]
     public sealed class EmagComponent : Component
     {
-        [DataField("maxCharges")]
+        [DataField("maxCharges"), ViewVariables(VVAccess.ReadWrite)]
         public int MaxCharges = 3;
 
-        [DataField("charges")]
+        [DataField("charges"), ViewVariables(VVAccess.ReadWrite)]
         public int Charges = 3;
 
-        [DataField("rechargeTime")]
+        [DataField("rechargeTime"), ViewVariables(VVAccess.ReadWrite)]
         public float RechargeTime = 90f;
         public float Accumulator = 0f;
     }
