+++ Content.Shared/Interaction/IInteractUsing.cs
@@ -71,13 +71,7 @@ namespace Content.Shared.Interaction
         /// </summary>
         public EntityCoordinates ClickLocation { get; }
 
-        /// <summary>
-        ///     If true, this prediction is also being predicted client-side. So care has to be taken to avoid audio
-        ///     duplication.
-        /// </summary>
-        public bool Predicted { get; }
-
-        public InteractUsingEvent(EntityUid user, EntityUid used, EntityUid target, EntityCoordinates clickLocation, bool predicted = false)
+        public InteractUsingEvent(EntityUid user, EntityUid used, EntityUid target, EntityCoordinates clickLocation)
         {
             // Interact using should not have the same used and target.
             // That should be a use-in-hand event instead.
@@ -88,7 +82,6 @@ namespace Content.Shared.Interaction
             Used = used;
             Target = target;
             ClickLocation = clickLocation;
-            Predicted = predicted;
         }
     }
 }
